16th of March, 2024 (12:20 PM EST) Update
== general ==
::Left4Bot's Survivors, Specials, Tanks, and Witches tables are turned into arrays. All code that concerns them and iterates over them are changed accordingly.

Errors could still abound -- more testing needs to be done.

== left4bots.nut ==

Clear L4D1Survivors table in AddonStop.

Optimized IsHandledSurvivor, IsHandledBot, IsHandledL4D1Bot into a *single* line of code.

Updated entity checking in a HELLUVA ton entity iteration functions to CHECK for boolean and integer values in those entities first before moving onto more intense calculations.

CountOtherStandingSurvivorsWithin partially optimized to increment with an integerized boolean value.

== left4bots_events.nut ==

Updated OnPostPlayerSpawn to use params["userid"] as, well, another parameter. Makes using user ids to initialize stuff easier for players.

On the subject of that, SuvivorFlow is greatly optimized here -- instead of being updated *every* tick in OnThinker, the value (as a table) is instead initialized with the player spawning and being replaced with a bot (and vice versa for the latter,) and the *relevant* values -- those being whether a survivor is in a checkpoint or not, and how much "flow" they have in the map -- being updated. (Defining and then redefining a table per game tick doesn't do wonders for performance.)

Also speaking of which, OnThinker is slightly optimized to use a few less branches.

Greatly optimized OnGameEvent_player_death. For one thing, a few debug-only stuff is sectioned off into its own, well, debug function. For the other thing, the code in the hook is simplified just a tad.

Commented out some "if" conditions throughout the script that checks if, say, a player or user id is in a table before deleting them -- according to an older Squirrel manual, the delete keyword does nothing if that key is not in the table, so this is a micro-optimization.
(With that said, this doesn't change how array values have to be removed.)

Replaced certain instances of "GetPlayerUserId()" in event hooks with the user ids used in event parameters, since using the former is redundant.